pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tacc"
version = "0.1.0"
description = "Time As Computation Cost: computational-capacity model of time dilation"
authors = [{name = "Robert Brodie", email = "robbybrodie@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "pandas>=1.3.0",
    "pyyaml>=5.4.0",
    "networkx>=2.6.0",
    "scikit-learn>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "jupyter>=1.0.0",
    "notebook>=6.0.0"
]

[project.urls]
"Homepage" = "https://github.com/robbybrodie/time_as_computation_cost"
"Repository" = "https://github.com/robbybrodie/time_as_computation_cost.git"
"Issues" = "https://github.com/robbybrodie/time_as_computation_cost/issues"

[project.scripts]
tacc = "tacc.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
